+++ test/utils/random.h
@@ -7,6 +7,11 @@
 extern "C" {
 #endif
 
+/**
+ * @brief Fill \p buffer with random data.
+ * @param[out] buffer   Buffer to fill.
+ * @param[in] size      Buffer size.
+ */
 void test_random(void* buffer, size_t size);
 
 #ifdef __cplusplus